+++ src/Gaffer/FileSequencePathFilter.cpp
@@ -75,7 +75,7 @@ void FileSequencePathFilter::setMode( Keep mode )
 	changedSignal()( this );
 }
 
-void FileSequencePathFilter::doFilter( std::vector<PathPtr> &paths ) const
+void FileSequencePathFilter::doFilter( std::vector<PathPtr> &paths, const IECore::Canceller *canceller ) const
 {
 	paths.erase(
 		std::remove_if(