+++ Changes.md
@@ -38,6 +38,7 @@ API
 - Path : Added optional `canceller` argument to `isValid()`, `isLeaf()`, `propertyNames()`, `property()` and `children()` methods.
 - ScenePath : Added support for cancellation of calls to `isValid()` and `children()`.
 - FileSystemPath : Added support for cancellation of calls to `children()` and `property()`.
+- PathFilter : Added support for cancellation of calls to `filter()`.
 
 Breaking Changes
 ----------------
@@ -58,6 +59,7 @@ Breaking Changes
   - Renamed `Type` enum to `Interpolation`.
   - Renamed Key's `set/getType()` accessors to `set/getInterpolation()`.
 - Path : Added `canceller` arguments to virtual methods. Note that Python subclasses can be made compatible with both Gaffer 0.60 and 0.61 simply by adding a `canceller = None` argument.
+- PathFilter : Added `canceller` argument to `doFilter()` method. Note that Python subclasses can be made compatible with both Gaffer 0.60 and 0.61 simply by adding a `canceller = None` argument.
 
 0.60.7.0 (relative to 0.60.6.1)
 ========